requirements.txt
amqp==2.1.1
anyjson==0.3.3
billiard==3.5.0.2
celery==4.0.0
cffi==1.9.1
cryptography==1.6
decorator==4.0.10
Django==1.10.3
django-celery==3.1.17
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-rest-knox==2.2.0
djangorestframework==3.5.3
idna==2.1
ipython==5.1.0
ipython-genutils==0.1.0
kombu==4.0.0
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.17
Pygments==2.1.3
pyOpenSSL==16.2.0
python-memcached==1.58
pytz==2016.7
requests==2.12.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
